+++ src/plonemeeting/restapi/services/get.py
@@ -1,94 +1,81 @@
 # -*- coding: utf-8 -*-
 
-from imio.helpers.content import uuidsToObjects
 from plone import api
-from plone.restapi.interfaces import ISerializeToJson
-from plone.restapi.interfaces import ISerializeToJsonSummary
-from plone.restapi.services.content.get import ContentGet
-from plonemeeting.restapi.utils import use_obj_serializer
+from plonemeeting.restapi.config import CONFIG_ID_NOT_FOUND_ERROR
+from plonemeeting.restapi.services.search import BaseSearchGet
+from plonemeeting.restapi.utils import rest_uuid_to_object
 from zExceptions import BadRequest
-from zope.component import queryMultiAdapter
 
 
 UID_REQUIRED_ERROR = 'The "UID or uid" parameter must be given!'
-UID_NOT_FOUND_ERROR = 'No element found with UID "%s"!'
-UID_NOT_ACCESSIBLE_ERROR = (
-    'Element with UID "%s" was found but user "%s" can not access it!'
-)
 UID_WRONG_TYPE_ERROR = (
     "The element UID does not correspond to the type managed by this endpoint! "
     "Consider using @get endpoint or another specific endpoint!"
 )
 
 
-def _get_obj_from_uid(uid):
-    """ """
-    # change self.context with element found with self.uid
-    objs = uuidsToObjects(uuids=uid)
-    if not objs:
-        # try to get it unrestricted
-        objs = uuidsToObjects(uuids=uid, unrestricted=True)
-        if objs:
-            raise BadRequest(
-                UID_NOT_ACCESSIBLE_ERROR
-                % (uid, api.user.get_current().getId())
-            )
-        else:
-            raise BadRequest(UID_NOT_FOUND_ERROR % uid)
-    return objs[0]
-
-
-class UidGet(ContentGet):
+class UidSearchGet(BaseSearchGet):
     """Returns a serialized content object based on required UID parameter."""
 
+    required_meta_type_id = None
+
     def __init__(self, context, request):
-        super(UidGet, self).__init__(context, request)
+        super(UidSearchGet, self).__init__(context, request)
         self.uid = self._uid
+        self.config_id = self._config_id
+        if self.config_id:
+            self.tool = api.portal.get_tool("portal_plonemeeting")
+            self.cfg = self.tool.get(self.config_id, None)
+            if not self.cfg:
+                raise BadRequest(CONFIG_ID_NOT_FOUND_ERROR % self.config_id)
+
+    @property
+    def _config_id(self):
+        return super(UidSearchGet, self)._config_id
 
     @property
     def _uid(self):
-        if "UID" not in self.request.form and "uid" not in self.request.form:
+        uid = self.request.form.get("UID") or self.request.form.get("uid")
+        if not uid:
             raise BadRequest(UID_REQUIRED_ERROR)
-        return self.request.form.get("UID") or self.request.form.get("uid")
+        return uid
 
-    def _check_obj(self):
+    def _check_res_type(self, res):
         """ """
-        return
+        if self.required_meta_type_id and res["items"]:
+            # we have the portal_type in res, we need to get the meta_type
+            # to compare it with self.required_meta_type_id
+            portal_type = api.portal.get_tool("portal_types")[res["items"][0]["@type"]]
+            # AT, meta_type is directly stored in content_meta_type
+            if portal_type.__class__.__name__ == 'DynamicViewTypeInformation':
+                meta_type = portal_type.content_meta_type
+            else:
+                # DX, klass is the dotted path the python klass
+                meta_type = portal_type.klass.split('.')[-1]
+            if self.required_meta_type_id != meta_type:
+                raise BadRequest(UID_WRONG_TYPE_ERROR)
 
     def reply(self):
-        obj = _get_obj_from_uid(self.uid)
-        self.context = obj
-        self._check_obj()
-
-        fullobjects = use_obj_serializer(self.request.form)
-        if fullobjects:
-            serializer = queryMultiAdapter((self.context, self.request), ISerializeToJson)
+        """ """
+        res = super(UidSearchGet, self).reply()
+        self._check_res_type(res)
+        if res["items"]:
+            # we only have one single result
+            res = res["items"][0]
         else:
-            serializer = queryMultiAdapter((self.context, self.request), ISerializeToJsonSummary)
-
-        if serializer is None:
-            self.request.response.setStatus(501)
-            return dict(error=dict(message="No serializer available."))
-
-        if fullobjects:
-            return serializer(version=self.request.get("version"))
-        else:
-            return serializer()
+            # will raise if element exist but inaccessible or not exist
+            # do not try_restricted as it was just done in this endpoint
+            rest_uuid_to_object(self.uid, try_restricted=False, in_name_of=self.in_name_of)
+        return res
 
 
-class ItemGet(UidGet):
+class ItemGet(UidSearchGet):
     """Get an item from it's UID, just check that returned value is a MeetingItem."""
 
-    def _check_obj(self):
-        """ """
-        if not self.context.__class__.__name__ == "MeetingItem":
-            raise BadRequest(UID_WRONG_TYPE_ERROR)
+    required_meta_type_id = "MeetingItem"
 
 
-class MeetingGet(UidGet):
+class MeetingGet(UidSearchGet):
     """Get a meeting from it's UID, just check that returned value is a Meeting."""
 
-    def _check_obj(self):
-        """ """
-        if not self.context.__class__.__name__ == "Meeting":
-            raise BadRequest(UID_WRONG_TYPE_ERROR)
+    required_meta_type_id = "Meeting"