+++ src/plonemeeting/restapi/tests/test_services_get.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from DateTime import DateTime
 from datetime import datetime
+from DateTime import DateTime
+from plonemeeting.restapi.config import IN_NAME_OF_UNAUTHORIZED
 from plonemeeting.restapi.serializer.meeting import HAS_MEETING_DX
-from plonemeeting.restapi.services.get import UID_NOT_ACCESSIBLE_ERROR
-from plonemeeting.restapi.services.get import UID_NOT_FOUND_ERROR
 from plonemeeting.restapi.services.get import UID_REQUIRED_ERROR
 from plonemeeting.restapi.services.get import UID_WRONG_TYPE_ERROR
 from plonemeeting.restapi.tests.base import BaseTestCase
+from plonemeeting.restapi.utils import IN_NAME_OF_CONFIG_ID_ERROR
+from plonemeeting.restapi.utils import UID_NOT_ACCESSIBLE_ERROR
+from plonemeeting.restapi.utils import UID_NOT_FOUND_ERROR
 from Products.PloneMeeting.tests.PloneMeetingTestCase import DEFAULT_USER_PASSWORD
 
 import transaction
@@ -149,6 +151,45 @@ class testServiceGetUid(BaseTestCase):
         self.assertTrue("extra_include_proposing_group" in json)
         self.assertTrue("externalIdentifier" in json)
         self.assertTrue("itemReference" in json)
+
+    def test_restapi_get_uid_in_name_of(self):
+        """Check when using parameter in_name_of"""
+        endpoint_url = "{0}/@get?UID={1}&in_name_of=pmCreator1".format(
+            self.portal_url, self.item1_uid
+        )
+        response = self.api_session.get(endpoint_url)
+        # config_id is required when using in_name_of
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(), {u"message": IN_NAME_OF_CONFIG_ID_ERROR, u"type": u"BadRequest"}
+        )
+        # with config_id the element is correctly returned
+        endpoint_url += "&config_id=%s" % self.meetingConfig.getId()
+        response = self.api_session.get(endpoint_url)
+        self.assertEqual(response.status_code, 200, response.content)
+        self.assertEqual(response.json()["UID"], self.item1_uid)
+        # try to get an inaccessible element
+        endpoint_url = endpoint_url.replace(self.item1_uid, self.item2_uid)
+        response = self.api_session.get(endpoint_url)
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            response.json(),
+            {
+                u"message": UID_NOT_ACCESSIBLE_ERROR % (self.item2_uid, "pmCreator1"),
+                u"type": u"BadRequest",
+            },
+        )
+        # must be MeetingManager to use in_name_of
+        self.api_session.auth = ("pmCreator1", DEFAULT_USER_PASSWORD)
+        response = self.api_session.get(endpoint_url)
+        self.assertEqual(response.status_code, 401)
+        self.assertEqual(
+            response.json(),
+            {
+                u"message": IN_NAME_OF_UNAUTHORIZED % "pmCreator1",
+                u"type": u"Unauthorized",
+            },
+        )
 
     def test_restapi_get_uid_extra_include_pod_templates(self):
         """Test the extra_include=pod_templates."""